Added an abstract class above AbstractMLP.
